requirements.txt
azure-identity
azure-keyvault
azure-storage-blob
openai
